ci: add gulp e2e test (#3318)
* Add E2E test to validate Gulp with Yarn

Adds an E2E test for coverage and validation that Yarn berry is able to install gulp successfully and then verify it can be used successfully.

After installing gulp, a simple script is added to create a default gulp task that prints a statement. Running the gulp command runs the task, making sure the bare basic functionality is able to run.

Then, in order to make sure that we are able to import and use the gulp api, a test  is used to use the src function from gulp to load the script file that's running.
Also, it makes sure we can use gulp with a specific task name as well.

This should be a start to verifying gulp is installable with yarn and working as intended.

* Shorten & Cleanup E2E scripts

* Corrected filename to gulpfile.js so gulp will find it by default

* Update e2e-gulp-workflow.yml

* Add Gulp E2E badge to Readme Toolchains

Adds badge to display the status of Gulp E2E tests to the ReadMe alongside the rest.

* Update PlugnPlay.md to mention support for Gulp

Gulp is successfully running with Yarn berry from version 4 and up. This adds it to the compatibility list.
Yarn may potentially support earlier versions, but the E2E tests succeed for the 4 line, and not 3.91 and below due to varying issues.

Co-authored-by: Maël Nison <nison.mael@gmail.com>