Updates the audit implementation to the bulk endpoint (#5501)
**What's the problem this PR addresses?**

The current audit implementation uses the older `/audit/quick` endpoint,
which has various problems. One particular is that its design requires
to submit a nested payload, but since it doesn't make much sense in our
case (because most of Yarn installs are flat), we flatten the package
list. It causes problems when multiple packages with different versions
can be found in the tree.

Fixes #3861
Fixes #4117 
Fixes #5408

Closes #5409 (Supercedes it)

---
Edit by @merceyz 

Fixes #5450
Fixes #2507
Fixes #3778
Fixes #3945
Closes #5309 (Doesn't have a reproduction so I'm assuming it's the same
as the others)

---

**How did you fix it?**

This change rewrites `yarn npm audit` to use the new endpoint. As part
of the migration a couple of fields are reworked (`Via` is replaced by
`Dependents`, the versions are now part of a tree item rather than
concatenated, we don't get the "recommendation" anymore). The options
remain the same for now.

It's possible that some registries don't support the bulk endpoint.
Given that it's fairly straightforward to implement, that it's been
released for some time now, and that without it we would end up with an
invalid `audit` implementation, I'd tend to let them deal with that.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.