Fixes merge conflict resolution always querying the network (#5504)
**What's the problem this PR addresses?**

I broke part of the automatic conflict resolution in #5102 - a code
supposed to make Yarn update the lockfile metadata when merging
incompatible lockfile versions was incorrectly causing Yarn to do that
for each merge conflict resolution. It can be very frustrating when the
project includes some dependencies that cannot be easily pulled from the
registry (for example when they are a npm git repo which doesn't build
anymore due to ERESOLVE 😄).

**How did you fix it?**

Fix the code to remove the `0` from `Math.min(0, ...)` which was
obviously causing the version to always be `0`.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.