Implements MOTD mechanism (#5530)
**What's the problem this PR addresses?**

It can be difficult to know some of the less-obvious features in Yarn.
Many people, even amongst our power users, aren't aware of some
particular things that Yarn allows them to do.

**How did you fix it?**

A [MOTD](https://en.wikipedia.org/wiki/Message_of_the_day) mechanism
will show a tip for each `yarn install` command run after 8am local TZ.
It'll also check whether the current Yarn version is outdated, and
replace the MOTD by an upgrade prompt in that case.

This mechanism only affects `yarn install` at the moment - perhaps it
should be also applied to `yarn add`, since zero-install users may not
use `install` as much, but our current philosophy has been to put
similar extra messaging inside `install` so far, so I preferred to keep
it that way.

The MOTD relies on the telemetry code (in part because it's convenient,
but also because the MOTD messages are pulled from the network, so in a
way they could lead to telemetry, if our backend was configured to track
requests made to it). They can be disabled by either disabling telemetry
altogether, or disabling MOTD specifically via `enableMotd: false`.

The https://repo.yarnpkg.com/tags endpoint (used to both get the latest
version and the motd, to avoid having to make two requests) has a 10mns
cache duration on Cloudflare. The MOTD are currently hardcoded within
the Cloudflare worker, but I'll update it to pull the list from this
repository after the PR lands.

If the request fails, the MOTD is skipped. I didn't add a timeout. I
considered improving `StreamReport` to allow pausing the display (while
the execution would continue) until the request would return, but it
felt a little over-engineered - the MOTD code is only executed once a
day, and never on CI. It'd also have required to either make this
feature exclusive to `startTimerPromise`, or to make `hasError` /
`exitCode` asynchronous, which was outside the scope of this work.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.